Moved resetChildLanes into complete work
This enabled us to remove a few hot path tag-type checks, but did not otherwise change any functionality.
